Move test to rest_api.rs
